feat: Adds custom reporter capability
